pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "asyncio>=4.0.0",
    "discord-py>=2.6.3",
    "pynacl>=1.5.0",
    "requests>=2.32.5",
    "yt-dlp>=2025.8.27",
]
